Accept ServicePlugin args and args as data
This patch adds several arguments to the module. Most of these are
parameters to ServicePlugin. So many plugins inherit from this plugin
it seems worth offering some special support for it.

The notable one is 'data'. I've typed this 'raw' which allows us to
pass an arbitrary data structure through it. This is useful because
some ServicePlugin arguments are compound, EG file_dir_names. It's
also useful for passing arguments other than those ServicePlugin ones
I added.

Depends-On: I77cbcac754fa3ecfe978bc9d2b642392b6bcea94